switch back to using build script
apparently when running PHD on CI, the `phd-runner` binary isn't
in the Propolis workspace, so we need to determine the Git rev at build
time, instead.
