+++ phd-tests/runner/Cargo.toml
@@ -13,7 +13,6 @@ doctest = false
 anyhow.workspace = true
 backtrace.workspace = true
 camino.workspace = true
-cargo_metadata.workspace = true
 clap = { workspace = true, features = ["derive"] }
 phd-framework.workspace = true
 phd-tests.workspace = true
@@ -22,3 +21,7 @@ tracing-appender.workspace = true
 tracing-bunyan-formatter.workspace = true
 tracing-subscriber = { workspace = true, features = ["env-filter"] }
 uuid.workspace = true
+
+[build-dependencies]
+anyhow.workspace = true
+cargo_metadata.workspace = true
