+++ phd-tests/runner/src/config.rs
@@ -2,7 +2,6 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at https://mozilla.org/MPL/2.0/.
 
-use anyhow::Context;
 use camino::Utf8PathBuf;
 use clap::{Args, Parser, Subcommand};
 use phd_framework::server_log_mode::ServerLogMode;
@@ -141,65 +140,16 @@ impl RunOptions {
             )));
         }
 
-        // Otherwise, determine the Git SHA of the workspace's Cargo git dep on
+        // Otherwise, use the Git revision of the workspace's Cargo git dep on
         // crucible-upstairs, and use the same revision for the downstairs
         // binary artifact.
-        fn extract_crucible_dep_sha(
-            src: &cargo_metadata::Source,
-        ) -> anyhow::Result<&str> {
-            const CRUCIBLE_REPO: &str =
-                "https://github.com/oxidecomputer/crucible";
-
-            let src = src.repr.strip_prefix("git+").ok_or_else(|| {
-                anyhow::anyhow!("Crucible package's source should be from git")
-            })?;
-
-            if !src.starts_with(CRUCIBLE_REPO) {
-                println!("cargo:warning=expected Crucible package's source to be {CRUCIBLE_REPO:?}, but is {src:?}");
-            }
-
-            let rev = src.split("?rev=").nth(1).ok_or_else(|| {
-                anyhow::anyhow!(
-                    "Crucible package's source should have a revision"
-                )
-            })?;
-            let mut parts = rev.split('#');
-            let sha = parts.next().ok_or_else(|| {
-                anyhow::anyhow!(
-                    "Crucible package's source should have a revision"
-                )
-            })?;
-            assert_eq!(Some(sha), parts.next());
-            Ok(sha)
-        }
-
-        let metadata = cargo_metadata::MetadataCommand::new()
-            .exec()
-            .context("Failed to get cargo metadata")?;
-
-        let crucible_pkg = metadata
-            .packages
-            .iter()
-            .find(|pkg| pkg.name == "crucible")
-            .ok_or_else(|| {
-                anyhow::anyhow!(
-                    "Failed to find Crucible package in cargo metadata"
-                )
-            })?;
-
-        let crucible_src = crucible_pkg.source.as_ref().ok_or_else(|| {
-                anyhow::anyhow!("Crucible package should not be a workspace member, and therefore should have source metadata")
-            })?;
-
-        let crucible_sha =
-            extract_crucible_dep_sha(crucible_src).with_context(|| {
-                format!(
-                    "Failed to extract Crucible source SHA from {crucible_src:?}"
-                )
-            })?;
-
+        //
+        // The Git revision of Crucible we depend on is determined when building
+        // `phd-runner` by the build script, so that the `phd-runner` binary can
+        // be run even after moving it out of the Propolis cargo workspace.
+        let crucible_git_rev = env!("PHD_CRUCIBLE_GIT_REV");
         Ok(Some(phd_framework::CrucibleDownstairsSource::BuildomatGitRev(
-            crucible_sha.to_string(),
+            crucible_git_rev.to_string(),
         )))
     }
 }
